+++ NEWS.md
@@ -30,6 +30,8 @@ NEW FUNCTIONS
 
 * `data_unite()`, to merge values of multiple variables into one new variable.
 
+* `data_modify()`, to create new variables in a data frame.
+
 BUG FIXES
 
 * `standardize_parameters()` now works when the package namespace is in the model